// src/types/PalDetails.ts
export interface PalDetails {
  id: string;
  name: string;
  internalName: string;
  deName: string;
  enName: string;
  esMxName: string;
  esName: string;
  frName: string;
  idName: string;
  itName: string;
  koName: string;
  plName: string;
  ptBrName: string;
  ruName: string;
  thName: string;
  trName: string;
  viName: string;
  zhHansName: string;
  zhHantName: string;
  jaName: string;
  breedingPower: number;
  probabilityMale: number;
  probabilityFemale: number;
  image: string;
}